% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserting_functions.R
\name{allowed_vehicle_details_variables}
\alias{allowed_vehicle_details_variables}
\title{Function to return allowed variables for the \code{`vehicle_details`} table
in an vehicle emissions database.}
\usage{
allowed_vehicle_details_variables()
}
\value{
Character vector.
}
\description{
Function to return allowed variables for the \code{`vehicle_details`} table
in an vehicle emissions database.
}
\author{
Stuart K. Grange
}
