% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_post_processor.R
\name{character_validity_to_binary}
\alias{character_validity_to_binary}
\title{Function to transform Opus's character-based RSD post-processsor validity 
codes to a binary validity.}
\usage{
character_validity_to_binary(x)
}
\arguments{
\item{x}{Opus's character-based RSD post-processsor validity vector.}
}
\value{
Logical vector.
}
\description{
Function to transform Opus's character-based RSD post-processsor validity 
codes to a binary validity.
}
\author{
Stuart K. Grange.
}
