% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_post_processor.R
\name{clean_opus_post_processor}
\alias{clean_opus_post_processor}
\title{Function to clean Opus's RSD post-processsor tibbles (tables) after reading
with \code{\link{read_opus_post_processor}}.}
\usage{
clean_opus_post_processor(df, df_images = NA)
}
\arguments{
\item{df}{Tibble/data frame from \code{\link{read_opus_post_processor}}.}

\item{df_images}{An additional tibble from 
\code{\link{read_opus_post_processor}} to join vehicle registrations to the 
table.}
}
\value{
Tibble.
}
\description{
Function to clean Opus's RSD post-processsor tibbles (tables) after reading
with \code{\link{read_opus_post_processor}}.
}
\seealso{
\code{\link{read_opus_post_processor}}
}
\author{
Stuart K. Grange
}
