% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{get_all_vehicle_makes}
\alias{get_all_vehicle_makes}
\title{Function to return all vehicle makes from a vehicle emissions database.}
\usage{
get_all_vehicle_makes(con)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}
}
\value{
Character vector.
}
\description{
Function to return all vehicle makes from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
