% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{get_registrations_for_makes}
\alias{get_registrations_for_makes}
\title{Function to return registrations for particular vehicle makes.}
\usage{
get_registrations_for_makes(con, make)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{make}{A vector of vehicle makes.}
}
\value{
Data frame.
}
\description{
Function to return registrations for particular vehicle makes.
}
\section{\code{\link{get_all_vehicle_makes}}}{
NA
}

