% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserting_functions.R
\name{get_sessions}
\alias{get_sessions}
\title{Function to return session values in the \code{`sessions`} table in a vehicle
emissions database.}
\usage{
get_sessions(con)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}
}
\value{
Character vector.
}
\description{
Function to return session values in the \code{`sessions`} table in a vehicle
emissions database.
}
\author{
Stuart K. Grange
}
