% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserting_functions.R
\name{get_sites}
\alias{get_sites}
\title{Function to return site values in the \code{`sites`} table in a vehicle
emissions database.}
\usage{
get_sites(con)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}
}
\value{
Character vector.
}
\description{
Function to return site values in the \code{`sites`} table in a vehicle
emissions database.
}
\author{
Stuart K. Grange
}
