% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_by_site}
\alias{import_by_site}
\title{Function to import vehicle emissions data by site from a vehicle emissions
database.}
\usage{
import_by_site(con, site = NA, parse_dates = TRUE, verbose = FALSE)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{site}{A vector of sites to return. If not used, data from all sites
will be selected.}

\item{parse_dates}{Should date variables be parsed?}

\item{verbose}{Should the function give messages?}
}
\value{
Data frame.
}
\description{
Function to import vehicle emissions data by site from a vehicle emissions
database.
}
\author{
Stuart K. Grange
}
