% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_distinct_registrations}
\alias{import_distinct_registrations}
\title{Function to import all distinct registrations from a vehicle emissions 
database.}
\usage{
import_distinct_registrations(con)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}
}
\value{
Data frame.
}
\description{
Function to import all distinct registrations from a vehicle emissions 
database.
}
\author{
Stuart K. Grange
}
