% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_meteorology}
\alias{import_meteorology}
\title{Function to import meteorological data from a vehicle emissions database.}
\usage{
import_meteorology(con, site = NA, spread = TRUE, verbose = FALSE)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{site}{A vector of sites.}

\item{spread}{Should the table be reshaped and made wider?}

\item{verbose}{Should the function give messages?}
}
\value{
Data frame.
}
\description{
Function to import meteorological data from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
