% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_sessions}
\alias{import_sessions}
\title{Function to import sessions data from a vehicle emissions database.}
\usage{
import_sessions(con)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}
}
\value{
Tibble.
}
\description{
Function to import sessions data from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
