% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_sites}
\alias{import_sites}
\title{Function to import sites data from a vehicle emissions database.}
\usage{
import_sites(con)
}
\arguments{
\item{con}{}
}
\value{
Data frame.
}
\description{
Function to import sites data from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
