% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_vehicle_details}
\alias{import_vehicle_details}
\title{Function to import vehicle detail data from a vehicle emissions database.}
\usage{
import_vehicle_details(con, registration = NA, spread = TRUE,
  verbose = FALSE)
}
\arguments{
\item{con}{}

\item{registration}{}

\item{spread}{}

\item{verbose}{}
}
\value{
Data frame.
}
\description{
Function to import vehicle detail data from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
