% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_vehicle_emissions}
\alias{import_vehicle_emissions}
\title{Function to import emissions data from a vehicle emissions database.}
\usage{
import_vehicle_emissions(con, registration = NA, verbose = FALSE)
}
\arguments{
\item{con}{}

\item{registration}{}

\item{verbose}{}
}
\value{
Data frame.
}
\description{
Function to import emissions data from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
