% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_vehicle_odometers}
\alias{import_vehicle_odometers}
\title{Function to import vehicle odometer data from a vehicle emissions database.}
\usage{
import_vehicle_odometers(con, registration = NA)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{registration}{A vector of vehicle registrations to filter return to.}
}
\value{
Data frame.
}
\description{
Function to import vehicle odometer data from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
