% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserting_functions.R
\name{insert_sessions}
\alias{insert_sessions}
\title{Function to insert session data into \code{`sessions`}.}
\usage{
insert_sessions(con, df)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{df}{Input data frame to insert into \code{`sessions`}.}
}
\value{
Invisible, a database insert.
}
\description{
Function to insert session data into \code{`sessions`}.
}
\author{
Stuart K. Grange
}
