% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserting_functions.R
\name{insert_vehicle_details}
\alias{insert_vehicle_details}
\title{Function to insert vehicle capture data into \code{`vehicle_details`}.}
\usage{
insert_vehicle_details(con, df, verbose = FALSE)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{df}{Input data frame to insert into \code{`vehicle_details`}.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible, a database insert.
}
\description{
Function to insert vehicle capture data into \code{`vehicle_details`}.
}
\author{
Stuart K. Grange
}
