% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutant_ratio_to_fuel_specific_emission.R
\name{pollutant_ratio_to_fuel_specific_emission}
\alias{pollutant_ratio_to_fuel_specific_emission}
\alias{read_molar_mass_table}
\alias{pollutant_to_co2_ratio}
\alias{nox_as_no2_equivalent}
\title{Functions to help calculate fuel specific emissions for various pollutants
using their ratio to CO[2].}
\usage{
pollutant_ratio_to_fuel_specific_emission(ratio_pollutant, pollutant,
  ratio_co, ratio_hc)

read_molar_mass_table()

pollutant_to_co2_ratio(pollutant, co2)

nox_as_no2_equivalent(no, no2)
}
\arguments{
\item{ratio_pollutant}{Ratio of pollutant to CO[2].}

\item{pollutant}{In \code{pollutant_ratio_to_fuel_specific_emission}, a 
string representing which polluant is being transformed, e.g. \code{"no"}. 
\code{pollutant_ratio_to_fuel_specific_emission} uses this string to find the
correct molar mass. In \code{pollutant_to_co2_ratio}, it is the variable 
which is to be transformed.}

\item{ratio_co}{Name of CO/CO[2] ratio variable.}

\item{ratio_hc}{Name of HC/CO[2] ratio variable.}

\item{co2}{Name of CO[2] ratio variable.}

\item{no}{Name of NO ratio variable.}

\item{no2}{Name of NO[2] ratio variable.}
}
\value{
Numeric vector or for \code{read_molar_mass_table}, a data frame.
}
\description{
Functions to help calculate fuel specific emissions for various pollutants
using their ratio to CO[2].
}
\author{
Stuart K. Grange
}
