% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_post_processor.R
\name{read_opus_post_processor}
\alias{read_opus_post_processor}
\title{Function to read Opus's RSD post-processsor data files.}
\usage{
read_opus_post_processor(file, verbose = FALSE)
}
\arguments{
\item{file}{Vector of file names.}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to read Opus's RSD post-processsor data files.
}
\author{
Stuart K. Grange
}
