% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_registration_images_table.R
\name{read_registration_images_table}
\alias{read_registration_images_table}
\title{Function to read files containing registration and image file data which act
as the linkage between the vehicle capture and vehicile details observational
units.}
\usage{
read_registration_images_table(file, distinct = TRUE)
}
\arguments{
\item{file}{Vector of file names.}

\item{distinct}{Should duplicated registration-image observations be removed?}
}
\value{
Tibble.
}
\description{
Function to read files containing registration and image file data which act
as the linkage between the vehicle capture and vehicile details observational
units.
}
\author{
Stuart K. Grange
}
