% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vehicle_details.R
\name{read_vehicle_details}
\alias{read_vehicle_details}
\title{Function to read joined vehicle data files.}
\usage{
read_vehicle_details(file, verbose = FALSE)
}
\arguments{
\item{file}{Vector of file names.}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to read joined vehicle data files.
}
\seealso{
\code{\link{tidy_odometer_table}}
}
\author{
Stuart K. Grange
}
