% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{sample_registrations}
\alias{sample_registrations}
\title{Function to return random registrations from a vehicle emissions database.}
\usage{
sample_registrations(con, n = 1)
}
\arguments{
\item{con}{}

\item{n}{}
}
\value{
Character vector.
}
\description{
Function to return random registrations from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
