% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_odometer_table.R
\name{tidy_odometer_table}
\alias{tidy_odometer_table}
\title{Function to tidy the messy odometer data contained in joined vehicle data 
files.}
\usage{
tidy_odometer_table(df, variables = 3, groups = 5)
}
\arguments{
\item{df}{Input tibble/data frame, usually from 
\code{\link{read_vehicle_details}}.}

\item{variables}{Number of variables in messy data.}

\item{groups}{Number of groups in messy data.}
}
\value{
Tibble.
}
\description{
Function to tidy the messy odometer data contained in joined vehicle data 
files.
}
\seealso{
\code{\link{read_vehicle_details}}
}
\author{
Stuart K. Grange
}
