% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dynamic_variables.R
\name{update_dynamic_variables}
\alias{update_dynamic_variables}
\title{Function to update dynamic variables in a vehicle emissions database.}
\usage{
update_dynamic_variables(con, verbose = FALSE)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible \code{con}.
}
\description{
Function to update dynamic variables in a vehicle emissions database.
}
\author{
Stuart K. Grange
}
