% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_functions.R
\name{update_session_counts}
\alias{update_session_counts}
\title{Function to update \code{`sites`}'s \code{session_count} variable in a
vehicle emissions database.}
\usage{
update_session_counts(con)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}
}
\value{
Invisible, a database insert.
}
\description{
Function to update \code{`sites`}'s \code{session_count} variable in a
vehicle emissions database.
}
\author{
Stuart K. Grange
}
