% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_functions.R
\name{vehicle_details_check}
\alias{vehicle_details_check}
\alias{vehicle_captures_check}
\alias{priority_vehicle_details_variables}
\alias{priority_vehicle_capture_variables}
\title{Functions to check variables within a data frame for a vehicle emissions 
database.}
\usage{
vehicle_details_check(df)

vehicle_captures_check(df)

priority_vehicle_details_variables(sort = TRUE)

priority_vehicle_capture_variables(sort = TRUE)
}
\arguments{
\item{df}{Data frame to test.}

\item{sort}{Should the variables be sorted alphabetically?}
}
\value{
Data frame with one observation.
}
\description{
Functions to check variables within a data frame for a vehicle emissions 
database.
}
