% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_by_session}
\alias{import_by_session}
\title{Function to import vehicle emissions data by session from a vehicle emissions
database.}
\usage{
import_by_session(con, session = NA, parse_dates = TRUE,
  valid_only = FALSE, verbose = FALSE)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{session}{A vector of sessions to return. If not used, data from all
sessions will be selected.}

\item{parse_dates}{Should date variables be parsed?}

\item{valid_only}{Should only valid data be returned?}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to import vehicle emissions data by session from a vehicle emissions
database.
}
\author{
Stuart K. Grange
}
