% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_meteorology}
\alias{import_meteorology}
\title{Function to import meteorological data from a vehicle emissions database.}
\usage{
import_meteorology(con, site = NA, spread = TRUE, verbose = FALSE)
}
\arguments{
\item{con}{}

\item{spread}{}

\item{verbose}{}
}
\value{
Data frame.
}
\description{
Function to import meteorological data from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
