% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_vehicle_captures}
\alias{import_vehicle_captures}
\title{Function to import vehicle capture data from a vehicle emissions database.}
\usage{
import_vehicle_captures(con, registration = NA, data_source = NA,
  valid_only = FALSE, spread = TRUE, verbose = FALSE)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{registration}{A vector of registrations to return. If not used, all
registrations will be selected.}

\item{data_source}{A vector of vehicle captures data sources to return.}

\item{valid_only}{Should only valid data be returned?}

\item{spread}{Should the table be reshaped and made wider?}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to import vehicle capture data from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
