% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_vehicle_details}
\alias{import_vehicle_details}
\title{Function to import vehicle detail data from a vehicle emissions database.}
\usage{
import_vehicle_details(con, registration = NA, spread = TRUE,
  parse_dates = TRUE, verbose = FALSE)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{registration}{A vector of registrations to return. If not used, all
registrations will be selected.}

\item{spread}{Should the table be reshaped and made wider?}

\item{parse_dates}{Should date variables be parsed?}

\item{verbose}{Should the function give messages?}
}
\value{
Data frame.
}
\description{
Function to import vehicle detail data from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
