% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_vehicle_emissions}
\alias{import_vehicle_emissions}
\title{Function to import emissions data from a vehicle emissions database.}
\usage{
import_vehicle_emissions(con, registration = NA, verbose = FALSE)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{registration}{A vector of registrations to return. If not used, all
registrations will be selected.}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to import emissions data from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
