% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{sample_registrations}
\alias{sample_registrations}
\title{Function to return random registrations from a vehicle emissions database.}
\usage{
sample_registrations(con, n = 1, sort = FALSE)
}
\arguments{
\item{con}{Database connection to a vehicle emissions database.}

\item{n}{Number of registrations to return.}

\item{sort}{Should the vector be sorted in alphabetical order?}
}
\value{
Character vector.
}
\description{
Function to return random registrations from a vehicle emissions database.
}
\author{
Stuart K. Grange
}
